+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-06-17 21:17:16
-// Commit: fix(storage): prevent double tap duplicate action in storage
+// Updated at: 2026-06-19 16:06:26
+// Commit: docs(storage): update readme and API contracts for storage
 
 class storageComponentState {
-  final String commitRef = '2090';
-  final String lastMessage = 'fix(storage): prevent double tap duplicate action in storage';
-  final DateTime timestamp = DateTime.parse('2026-06-17T21:17:16+05:00');
+  final String commitRef = '2136';
+  final String lastMessage = 'docs(storage): update readme and API contracts for storage';
+  final DateTime timestamp = DateTime.parse('2026-06-19T16:06:26+05:00');
 }